Moved bragg and c2_x definitions back to energy axis Moved energy definition back to 10-machine DCM object uses definition from energy axis Removed 12-energy since energy object moved back to 10-machine
